requirements.txt
rich~=12.2.0
GitPython~=3.1.27
semver~=2.13.0
